e2ee.h + qolmutils.* -> e2ee_common.*
qolmutils.* pair has been confusingly similar to qolmutility.*; on the
other hand, it's needed in quite a few places where e2ee.h used to be
needed so it made sense to merge the two and rename the pair to reflect
its meaning.
