+++ autotests/testkeyverification.cpp
@@ -8,7 +8,6 @@
 #include <QtCore/QDateTime>
 #include <e2ee/qolmaccount.h>
 #include "olm/sas.h"
-#include "e2ee/qolmutils.h"
 
 class TestKeyVerificationSession : public QObject
 {
