+++ lib/connection.h
@@ -22,7 +22,7 @@
 #include <functional>
 
 #ifdef Quotient_E2EE_ENABLED
-#include "e2ee/e2ee.h"
+#include "e2ee/e2ee_common.h"
 #include "e2ee/qolmoutboundsession.h"
 #include "keyverificationsession.h"
 #include "events/keyverificationevent.h"
