+++ lib/csapi/keys.h
@@ -7,7 +7,7 @@
 #include "csapi/definitions/cross_signing_key.h"
 #include "csapi/definitions/device_keys.h"
 
-#include "e2ee/e2ee.h"
+#include "e2ee/e2ee_common.h"
 
 #include "jobs/basejob.h"
