+++ lib/database.h
@@ -9,7 +9,7 @@
 
 #include <QtCore/QHash>
 
-#include "e2ee/e2ee.h"
+#include "e2ee/e2ee_common.h"
 
 namespace Quotient {
