+++ lib/e2ee/e2ee_common.h
@@ -47,6 +47,25 @@ inline bool isSupportedAlgorithm(const QString& algorithm)
 }
 
 #ifdef Quotient_E2EE_ENABLED
+
+#define QOLM_INTERNAL_ERROR_X(Message_, LastError_) \
+    qFatal("%s, internal error: %s", Message_, LastError_)
+
+#define QOLM_INTERNAL_ERROR(Message_) \
+    QOLM_INTERNAL_ERROR_X(Message_, lastError())
+
+#define QOLM_FAIL_OR_LOG_X(InternalCondition_, Message_, LastErrorText_)   \
+    do {                                                                   \
+        const QString errorMsg{ (Message_) };                              \
+        if (InternalCondition_)                                            \
+            QOLM_INTERNAL_ERROR_X(qPrintable(errorMsg), (LastErrorText_)); \
+        qWarning(E2EE).nospace() << errorMsg << ": " << (LastErrorText_);  \
+    } while (false) /* End of macro */
+
+#define QOLM_FAIL_OR_LOG(InternalFailureValue_, Message_)                      \
+    QOLM_FAIL_OR_LOG_X(lastErrorCode() == (InternalFailureValue_), (Message_), \
+                       lastError())
+
 struct Unencrypted {};
 struct Encrypted {
     QByteArray key;
@@ -56,7 +75,47 @@ using PicklingMode = std::variant<Unencrypted, Encrypted>;
 
 template <typename T>
 using QOlmExpected = Expected<T, OlmErrorCode>;
-#endif
+
+// Convert PicklingMode to key
+QUOTIENT_API QByteArray toKey(const PicklingMode &mode);
+
+//! \brief Initialise a buffer object for use with Olm calls
+//!
+//! Qt and Olm use different size types; this causes the warning noise
+QUOTIENT_API QByteArray bufferForOlm(size_t bufferSize);
+
+//! \brief Get a size of Qt container coerced to size_t
+//!
+//! It's a safe cast since size_t can easily accommodate the range between
+//! 0 and INTMAX - 1 that Qt containers support; yet compilers complain...
+inline size_t unsignedSize(auto qtBuffer)
+{
+    // The buffer size cannot be negative by definition, so this is deemed safe
+    return static_cast<size_t>(qtBuffer.size());
+}
+
+class QUOTIENT_API RandomBuffer : public QByteArray {
+public:
+    explicit RandomBuffer(size_t size);
+    ~RandomBuffer() { clear(); }
+
+    // NOLINTNEXTLINE(google-explicit-constructor)
+    QUO_IMPLICIT operator void*() { return data(); }
+    char* chars() { return data(); }
+    uint8_t* bytes() { return reinterpret_cast<uint8_t*>(data()); }
+
+    Q_DISABLE_COPY(RandomBuffer)
+    RandomBuffer(RandomBuffer&&) = default;
+    void operator=(RandomBuffer&&) = delete;
+};
+
+[[deprecated("Create RandomBuffer directly")]] //
+inline RandomBuffer getRandom(size_t bufferSize)
+{
+    return RandomBuffer(bufferSize);
+}
+
+#endif // Quotient_E2EE_ENABLED
 
 struct IdentityKeys
 {
