+++ lib/e2ee/qolmaccount.h
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "e2ee/e2ee.h"
+#include "e2ee/e2ee_common.h"
 #include "e2ee/qolmmessage.h"
 
 #include "csapi/keys.h"
