+++ lib/events/encryptedevent.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 #include "encryptedevent.h"
-#include "e2ee/e2ee.h"
+#include "e2ee/e2ee_common.h"
 #include "logging.h"
 
 using namespace Quotient;
